makes more cleanups/refactors to json
